// src/types/Experience.ts
export type Experience = {
    _id: string;
    order: number;
    logo: string;
    company: string;
    companyUrl?: string;
    position: string;
    location: string;
    duration: string;
    displayName?: string;
    altCompany?: string;
    altCompanyUrl?: string;
    team?: string;
    contract?: boolean;
    project?: string;
    projectUrl?: string;
    projectCompany?: string;
    projectCompanyUrl?: string;
    roundLogo?: boolean;
}